+++ src/yamk/__init__.py
@@ -1,3 +1,1 @@
-from packaging.version import Version
-
-__version__ = Version("6.0.1")
+__version__ = "6.0.1"