+++ src/yamk/lib.py
@@ -391,6 +391,23 @@ class CommandReport:
         )
 
 
+@dataclass(frozen=True, order=True)
+class Version:
+    major: int
+    minor: int
+    patch: int
+
+    @classmethod
+    def from_string(cls, version: str) -> Version:
+        version = re.sub("[^.0-9].*", "", version)
+        if version.endswith("."):
+            version += "0"
+        while version.count(".") < 2:
+            version += ".0"
+        major, minor, patch, *_ = map(int, version.split("."))
+        return cls(major, minor, patch)
+
+
 def flatten_vars(
     variables: dict[str, dict[str, Any]], base_dir: Path
 ) -> dict[str, Any]: