+++ docs/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+### Added
+
+- Added a version parser to avoid the packaging dependency
+
 ## [6.0.1] - 2024-03-31
 
 ### Added